// src/sdk/generated/models/BadgeCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BadgeCreate = {
  name: string;
  display_name: string;
  description?: (string | null);
  icon_url?: (string | null);
  color_code?: (string | null);
  rarity?: string;
  criteria: Record<string, any>;
  points_required?: (number | null);
  is_secret?: boolean;
};